Move webhook model/resource classes and tests to operator.webhooks
