+++ operator/src/main/java/oracle/kubernetes/operator/webhooks/model/AdmissionReview.java
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package oracle.kubernetes.operator.http.rest.model;
+package oracle.kubernetes.operator.webhooks.model;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
