+++ operator/src/main/java/oracle/kubernetes/operator/webhooks/model/Result.java
